\name{checkprogress}
\alias{checkprogress}
\title{Display a progress bar in a for-loop}
\description{
The function displays a text progress bar when running a for loop (wrapper for txtProgressBar function).
}
\usage{
checkprogress(endi, starti = 1, val = i)
}
\arguments{
\item{endi}{The (end) number of iterations in the for loop.}
\item{starti}{The number of the first iteration, usually 1.}
\item{val}{The counting variable in the for loop, e.g. i .}
}


